ai-summarizer-agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.0
google-cloud-aiplatform==1.42.1
pydantic==2.5.0
